:recycle: refactor biosample tasks test_on_failure
